singup + fix
